requirements.txt
aiohttp==3.9.5
feedparser==6.0.11
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.2.1
openai==1.40.2
python-telegram-bot==21.4
python-dotenv==1.0.1
